requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-dotenv
qrcode[pil]
python-multipart
psycopg2-binary
pydantic
email-validator
pydantic[email]
passlib[bcrypt]
minio
PyJWT
cryptography

